meeting-assistant-backend/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
openai
langchain
whisper
python-dotenv
groq